+++ lib/src/utils/asset_loader/asset_loader.dart
@@ -4,7 +4,6 @@ import 'asset_loader_io.dart'
     as platform;
 
 export 'asset_loader_base.dart';
-export 'asset_loader_flutter.dart';
 export 'asset_loader_io.dart' if (dart.library.js) 'asset_loader_web.dart';
 
 // ignore: unused_field